Create TCP_Client
File Transfer Protocol is a bit too much so this is a simple version with just 5 commands and these commands will be slightly different from the real ones.
Because it will be running the client and the server on the same machine, there should be two directories in the working directory. One called client and the other called server. These are the directories that files from the ftp commands will be stored.

The following commands can be implemented:

OPEN #
Attempts to open an FTP connection to 127.0.0.1 using the port number specified. Note that when testing your program the TAs will be using ports of their choosing.

GET filename
Requests a file from the server. This file will be loaded from the server directory, transferred by the server to the client, and saved by the client to the same name in the client directory.

PUT filename
Sends a file to the server. This file will be loaded from the client directory, transferred by the client to the server, and saved by the server to the same name in the server directory.

CLOSE
This closes the current connection to the server but keeps the client running so that the client can connect to another server if they wish to (and the TAs will wish to)

QUIT
Exits the client, closing any open connections